Rename commonMockData.ts to commonApplicationMockData.ts
Ticket: AT-6739
