+++ packages/api/portfolioDrafts/application/getApplicationStep.test.ts
@@ -3,7 +3,7 @@ import { ApiSuccessResponse, ErrorStatusCode, OtherErrorResponse, SuccessStatusC
 import { DATABASE_ERROR, NO_SUCH_APPLICATION_STEP, PATH_PARAMETER_REQUIRED_BUT_MISSING } from "../../utils/errors";
 import { DynamoDBDocumentClient, GetCommand, GetCommandOutput } from "@aws-sdk/lib-dynamodb";
 import { handler, NO_SUCH_PORTFOLIO_DRAFT } from "./getApplicationStep";
-import { mockApplicationStep } from "./commonMockData";
+import { mockApplicationStep } from "./commonApplicationMockData";
 import { mockClient } from "aws-sdk-client-mock";
 import { v4 as uuidv4 } from "uuid";
